vote receipt contents #203
