+++ src/test/java/org/rulez/demokracia/pdengine/VoteShowTest.java
@@ -1,8 +1,6 @@
 package org.rulez.demokracia.pdengine;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
-
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.Before;
